stock_price_movement_doc_commit
The research addresses the problem of discovering a model that can presage the rise and fall of 
stocks, as it is an arduous job as there may be myriad aspects that can impact share prices. This 
thesis finds an effective prediction model that yields an accurate prognosis with the most 
subordinate error percentage for both fundamental and social media data (Twitter).